+++ src/MythicScribe.ts
@@ -17,7 +17,7 @@ import { metaskillFileCompletionProvider } from './imports/completions/metaskill
 
 import { removeBracketsTextListener } from './imports/textchanges/bracketsremover';
 import { shortcutsProvider } from './imports/textchanges/shortcuts';
-import { loadGithubDatasets, updateDatasetMaps } from './objectInfos';
+import { loadGithubDatasets, updateDatasets } from './objectInfos';
 import { triggerfileCompletionProvider } from './imports/completions/triggerfileCompletionProvider';
 
 export let ctx: vscode.ExtensionContext;
@@ -42,7 +42,7 @@ export function activate(context: vscode.ExtensionContext) {
 		});	
 	}
 	else{
-		updateDatasetMaps();
+		updateDatasets();
 	}
 
 	// Config